server/requirements.txt
fastapi
uvicorn
motor
python-multipart
bcrypt
PyJWT